backend/requirements.txt
# Core
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
python-dotenv>=1.0.0

# Video processing
google-genai>=1.0.0
ffmpeg-python>=0.2.0

# WebSocket support (included in uvicorn[standard])
websockets>=12.0
